pyproject.toml
[tool.poetry]
name = "nianbot"
version = "2.0.0"
description = "The Discord Bot for the KyoStinV official Discord Server."
authors = ["gian <@0xgian>"]

[tool.poetry.dependencies]
python = "3.10.*"

"discord.py" = "2.1.0"

aiohttp = "3.8.3"
aiodns = "3.0.0"
beautifulsoup4 = "4.11.1"
tldextract = "3.4.0"
markdownify = "0.6.1"
psycopg2 = "2.9.5"

colorama = "0.4.6"
coloredlogs = "15.0.1"
emoji = "2.1.0"

arrow = "1.2.3"
more-itertools = "8.14.0"
python-dateutil = "2.8.2"
python-frontmatter = "1.0.0"
pyyaml = "6.0"
rapidfuzz = "2.11.1"
regex = "2022.9.13"
deepdiff = "5.8.1"

sentry-sdk = "1.9.10"
statsd = "3.3.0"
pydantic = "2.0.3"
pydantic-settings = "2.0.3"


[tool.poetry.dev-dependencies]
coverage = "7.3.1"
httpx = "0.25.0"
isort = "5.12.0"
pre-commit = "3.4.0"
pip-licenses = "4.3.3"
pytest = "7.4.2"
pytest-cov = "4.1.0"
pytest-subtests = "0.11.0"
pytest-xdist = "3.3.1"
ruff = "0.0.290"
taskipy = "1.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "python -m nianbot"
lint = "pre-commit run --all-files"
precommit = "pre-commit install"
build = "docker build -t ghcr.io/gentilinigm/nianbot:latest -f Dockerfile ."
push = "docker push ghcr.io/gentilinigm/nianbot:latest"
test-nocov = "pytest -n auto"
test = "pytest -n auto --cov-report= --cov --ff"
retest = "pytest -n auto --cov-report= --cov --lf"
html = "coverage html"
report = "coverage report"
isort = "isort ."

[tool.coverage.run]
branch = true
source_pkgs = ["nianbot"]
source = ["tests"]

[tool.isort]
multi_line_output = 6
order_by_type = false
case_sensitive = true
combine_as_imports = true
line_length = 120
atomic = true

[tool.ruff]
target-version = "py310"
extend-exclude = [".cache"]
ignore = [
    "ANN002", "ANN003", "ANN101", "ANN102", "ANN204", "ANN206", "ANN401",
    "B904",
    "C401", "C408",
    "D100", "D104", "D105", "D107", "D203", "D212", "D214", "D215", "D301",
    "D400", "D401", "D402", "D404", "D405", "D406", "D407", "D408", "D409", "D410", "D411", "D412", "D413", "D414", "D416", "D417",
    "E731",
    "RET504",
    "RUF005", "RUF012", "RUF015",
    "S311",
    "SIM102", "SIM108",
]
line-length = 120
select = ["ANN", "B", "C4", "D", "DTZ", "E", "F", "I", "ISC", "INT", "N", "PGH", "PIE", "Q", "RET", "RSE", "RUF", "S", "SIM", "T20", "TID", "UP", "W"]


[tool.ruff.per-file-ignores]
"tests/*" = ["ANN", "D"]

[tool.pytest.ini_options]
# This stops pytest from running functions named `setup` in test files.
addopts = "-p no:nose"
